+++ src/figrecipe/_params/_PLOTTING_METHODS.py
@@ -53,6 +53,7 @@ PLOTTING_METHODS = {
     "angle_spectrum",
     "magnitude_spectrum",
     "phase_spectrum",
+    "graph",
 }
 
 # EOF